完善高版本的自动更新功能

+++ updateapputils/src/main/java/update/UpdateAppUtils.kt
@@ -144,11 +144,13 @@ object UpdateAppUtils {
     fun deleteInstalledApk() {
         val apkPath = SPUtil.getString(DownloadAppUtils.KEY_OF_SP_APK_PATH, "")
         val appVersionCode = Utils.getAPPVersionCode()
-        val apkVersionCode = Utils.getApkVersionCode(apkPath)
+        val apkVersionCode = apkPath?.let { Utils.getApkVersionCode(it) }
         log("appVersionCode:$appVersionCode")
         log("apkVersionCode:$apkVersionCode")
-        (apkPath.isNotEmpty() && appVersionCode == apkVersionCode && apkVersionCode > 0).yes {
-            Utils.deleteFile(apkPath)
+        if (apkPath != null) {
+            (apkPath.isNotEmpty() && appVersionCode == apkVersionCode && apkVersionCode > 0).yes {
+                Utils.deleteFile(apkPath)
+            }
         }
     }
